// frontend/src/app/master/models/ZohoEmp.ts
export interface ZohoEmp {
  emp_Id: any;
  emp_Name: any;
  zohoEmp_Id: number;
  employee_Id: string;
  userName: string;
  designation: string;
  business_unit: string;
  status: string;
  yearOfExperience: string;
  mobile: string;
  emailID: string;
  dateOfJoining: string;
  department: string;
  employeeCategory: string;
  reporting_Manager: string;
  second_Reporting_To: string;
  work_location: string;
  entity: string;
  ownerName: string;
  recordId: string;
  approvalStatus: string;
  createdOn: string;     
  modifiedOn: string;   
}